rename NoProjects card to NoTokensStakedCard
