+++ frontend/src/lib/components/portfolio/NoTokensStakedCard.svelte
